+++ app/Pages/HomePage.tsx
@@ -8,7 +8,7 @@ import { useFocusEffect } from "@react-navigation/core";
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
 import Header from "../Components/Header";
 import DifficultySlider from "../Components/Home/DifficultySlider";
-import { getKeyString } from "../Functions/Auth0/token";
+import { getKeyString } from "../Functions/AsyncStorage/token";
 import { USERACTIVEGAMESBFFURL, USERGAMESTATISTICSBFFURL } from "@env";
 import { MaterialCommunityIcons } from "@expo/vector-icons";
 import Alert from "react-native-awesome-alerts";
