+++ app/Pages/StatisticsPage.tsx
@@ -4,7 +4,7 @@ import { Button, useTheme, ActivityIndicator } from "react-native-paper";
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
 import Header from "../Components/Header";
 import { useWindowDimensions } from "react-native";
-import { getKeyString } from "../Functions/Auth0/token";
+import { getKeyString } from "../Functions/AsyncStorage/token";
 import { USERGAMESTATISTICSBFFURL } from "@env";
 import { useNavigation } from "@react-navigation/native";
 import { PreferencesContext } from "../Contexts/PreferencesContext";
